overview lapjusik

+++ app/Http/Controllers/KemajuanProyekController.php
@@ -116,7 +116,6 @@ class KemajuanProyekController extends Controller
     public function viewInfo($id){
         $idPelaksanaan = Pelaksanaan::select('pelaksanaans.id')->where('proyek_id',$id)->get();
         $listInformasi = KemajuanProyek::select('kemajuan_proyeks.*')->whereIn('pelaksanaan_id',$idPelaksanaan)->get();
-        $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$id],['approvalStatus','=',0]])->first();
         $listPekerjaan = DB::table('jenis_pekerjaan')->select('jenis_pekerjaan.name')->where('proyek_id',$id)->get();
         $lizWork=array($listPekerjaan->count());
         //dd($listPekerjaan->count());
@@ -127,18 +126,7 @@ class KemajuanProyekController extends Controller
             //dd($counter);
         }
         //dd($lizWork);
-
-        if($pelaksanaan == null) {
-            DB::table('pelaksanaans')->insert([
-                'approvalStatus' => 0,
-                'createdDate' => now('GMT+7'),
-                'proyek_id' => $id,
-                'created_at' => now('GMT+7'),
-                'updated_at' => now('GMT+7')
-            ]);
-            $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$id],['approvalStatus','=',0]])->first();         
-        }
-            return view('listInformasi', compact('listInformasi','pelaksanaan','lizWork'));
+            return view('listInformasi', compact('listInformasi','lizWork'));
     }
 
     public function detailInfo($id) {
@@ -164,13 +152,13 @@ class KemajuanProyekController extends Controller
         return view('detailInformasi', compact('informasi','proyek','tanggal','foto','lizWork'));
     }
 
-    public function tambahInformasi($id){
-        $pelaksanaan = Pelaksanaan::find($id);
-        $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->whereIn('id',$pelaksanaan)->get();
-        //dd($idProyek);
+    public function tambahInformasi(){
+        $idProyek = Assignment::select('assignments.proyek_id')->where('pengguna_id',\Auth::user()->id)->get();
+        $proyekId = $idProyek[0]->proyek_id;
+        //dd($proyekId);
         $pekerjaan = DB::table('jenis_pekerjaan')->whereIn('proyek_id',$idProyek)->get();
         //dd($pekerjaan);
-        return view('tambahInformasi',compact('pelaksanaan','pekerjaan'));
+        return view('tambahInformasi',compact('pekerjaan','proyekId'));
     }
 
     public function tambahFoto($id){
@@ -209,10 +197,39 @@ class KemajuanProyekController extends Controller
     @param \Illuminate\Http\Request
     @return \Illuminate\Http\Response
     */
-    public function simpanInformasi($id, Request $request){
+    public function simpanInformasi(Request $request){
+        $idProyek = Assignment::select('assignments.proyek_id')->where('pengguna_id',\Auth::user()->id)->get();
+        $proyekId = $idProyek[0]->proyek_id;
 
-        $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->where('id',$id)->get();
-        $data = json_decode($idProyek);
+        //Bulan awal
+        $firstDate = DB::table('kemajuan_proyeks')->select('kemajuan_proyeks.reportDate')->min('reportDate');
+        $firstMonth = date('m', strtotime($firstDate));
+        $firstYear = date('Y', strtotime($firstDate));
+        
+        //Bulan dari Form
+        $inputMonth = date('m', strtotime($request->reportdate));
+        $inputYear = date('Y', strtotime($request->reportdate));
+
+        //Belum ada LAPJUSIK (Ubah -> Adjusted)
+        $adjustedMonth = $inputMonth - ($firstMonth-1);
+        //dd($adjustedMonth);
+
+        $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$proyekId],['approvalStatus','=',0],['bulan','=',$adjustedMonth]])->first();
+        //dd($pelaksanaan);
+
+        //Bikin LAPJUSIK baru
+        if($pelaksanaan == null) {
+
+            DB::table('pelaksanaans')->insert([
+                'approvalStatus' => 0,
+                'createdDate' => now('GMT+7'),
+                'bulan'=> $bulanNow,
+                'proyek_id' => $proyekId,
+                'created_at' => now('GMT+7'),
+                'updated_at' => now('GMT+7')
+            ]);
+            $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$proyekId],['approvalStatus','=',0]])->first();         
+        }
 
         $validator = Validator::make($request->all(),[
             'tipepekerjaan' => 'required',
@@ -229,7 +246,7 @@ class KemajuanProyekController extends Controller
             'tipeKemajuan' => $request->tipekemajuan,
             'value' => $request->nilai,
             'pekerjaan_id' => $request->tipepekerjaan,
-            'pelaksanaan_id' => $id,
+            'pelaksanaan_id' => $pelaksanaan->id,
             'created_at' => now('GMT+7'),
             'updated_at' => now('GMT+7')
         ]);
@@ -245,7 +262,7 @@ class KemajuanProyekController extends Controller
             'workCurrentValue' => $combinedValue
         ]);
 
-        $kemajuans = KemajuanProyek::select('kemajuan_proyeks.*')->where('pelaksanaan_id', $id)->get()->last();
+        $kemajuans = KemajuanProyek::select('kemajuan_proyeks.*')->where('pelaksanaan_id', $pelaksanaan->id)->get()->last();
         $kemajuan_id = $kemajuans->id;
 
         if ($request->file != null) {
@@ -268,7 +285,7 @@ class KemajuanProyekController extends Controller
         /*if($request->file!= null) {
             $path = $uploadedFile->store('/files');
         }*/
-        return redirect()->action('KemajuanProyekController@viewInfo',['id'=>$data[0]->proyek_id]);
+        return redirect()->action('KemajuanProyekController@viewInfo',['id'=>$proyekId]);
     }
 
     public function editInformasi($id){
@@ -276,27 +293,28 @@ class KemajuanProyekController extends Controller
         $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->whereIn('id',$idPelaksanaan)->get();
         $proyek = Proyek::find($idProyek[0]->proyek_id);
         $kemajuans = KemajuanProyek::find($id);
-
         $foto = DB::table('listPhoto')->select('listPhoto.*')->where('kemajuan_id',$id)->get();
-        //dd($foto);
-        return view('editInformasi', compact('kemajuans','proyek','foto'));
+        $pekerjaan = DB::table('jenis_pekerjaan')->where('proyek_id',$idProyek[0]->proyek_id)->get();
+        return view('editInformasi', compact('kemajuans','proyek','foto','pekerjaan'));
     }
 
     public function updateInformasi($id, Request $request){
-        
-        
-        $allId = DB::table('listPhoto')->select('listPhoto.id')->where('kemajuan_id',$id)->whereNotIn('id',$request->listId)->get();
-        $deletedId = json_decode($allId);
-        for($i=0;$i<sizeof($allId);$i++) {
+        //dd($request->listId);
+        if ($request->listId!=null) {
+            $allId = DB::table('listPhoto')->select('listPhoto.id')->where('kemajuan_id',$id)->whereNotIn('id',$request->listId)->get();
+            //dd($allId);
+            $deletedId = json_decode($allId);
+            //dd($deletedId);
+            for($i=0;$i<sizeof($allId);$i++) {
 
-            DB::table('listPhoto')->where('id',$deletedId[$i]->id)->delete();
+                DB::table('listPhoto')->where('id',$deletedId[$i]->id)->delete();
+            }
         }
         //dd($deletedId[0]->id);
         $idPelaksanaan = KemajuanProyek::select('kemajuan_proyeks.pelaksanaan_id')->where('id',$id)->get();
         $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->whereIn('id',$idPelaksanaan)->get();
         $data = json_decode($idProyek);
         $validator = Validator::make($request->all(),[
-            'description' => 'required',
             'reportDate' => 'required',
             'tipeKemajuan' => 'required',
             'value' => 'required',
@@ -305,11 +323,25 @@ class KemajuanProyekController extends Controller
     	]);
 
         $kemajuans = KemajuanProyek::find($id);
+        //dd($kemajuans);
+        $kemajuans->pekerjaan_id = $request->tipepekerjaan;
         $kemajuans->description = $request->description;
         $kemajuans->reportDate = $request->reportdate;
         $kemajuans->value = $request->nilai;
         $kemajuans->tipeKemajuan = $request->tipekemajuan;
         $kemajuans->save();
+        //dd($kemajuans->pekerjaan_id);
+
+        $realValue = DB::table('kemajuan_proyeks')->select('kemajuan_proyeks.value')->where('pekerjaan_id',$kemajuans->pekerjaan_id)->get();
+        //dd(json_decode($realValue));
+        $combinedValue = 0;
+        foreach($realValue as $value) {
+            $combinedValue += $value->value;
+        }
+
+        DB::table('jenis_pekerjaan')->where('id',$kemajuans->pekerjaan_id)->update([
+            'workCurrentValue' => $combinedValue
+        ]);
 
         return redirect()->action('KemajuanProyekController@viewInfo',['id'=>$data[0]->proyek_id]);
     }
@@ -319,17 +351,18 @@ class KemajuanProyekController extends Controller
         $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->whereIn('id',$idPelaksanaan)->get();
         $data = json_decode($idProyek);
         $kemajuans = KemajuanProyek::find($id);
-        dd($kemajuans);
+        //dd($kemajuans->pekerjaan_id);
         $kemajuans->delete();
 
-        $realValue = DB::table('kemajuan_proyeks')->select('kemajuan_proyeks.value')->where('pekerjaan_id',$request->tipepekerjaan)->get();
+        $realValue = DB::table('kemajuan_proyeks')->select('kemajuan_proyeks.value')->where('pekerjaan_id',$kemajuans->pekerjaan_id)->get();
         //dd(json_decode($realValue));
         $combinedValue = 0;
         foreach($realValue as $value) {
+            //dd($value->value);
             $combinedValue += $value->value;
         }
-
-        DB::table('jenis_pekerjaan')->where('id',$request->tipepekerjaan)->update([
+        //dd($combinedValue);
+        DB::table('jenis_pekerjaan')->where('id',$kemajuans->pekerjaan_id)->update([
             'workCurrentValue' => $combinedValue
         ]);
